CmdStan 2.31/2.32

+++ server/cache.js
@@ -62,15 +62,23 @@ function store_model(model_id, model) {
 }
 
 exports.delete_model = delete_model;
-function delete_model(model_id) {
-  // FIXME delete files too
+async function delete_model(model_id) {
+  if (!all_models.delete(model_id)) {
+    return false;
+  }
   let del_fits = [];
   for (let [fit_id, fit] of all_fits)
     if (fit.model_id === model_id)
       del_fits.push(fit_id);
   for (let fit_id of del_fits)
     delete_fit(fit_id);
-  return all_models.delete(model_id);
+  await write_models();
+  let model_dir = join(cache_dir, 'models', model_id);
+  for (let f of fs.readdirSync(model_dir)) {
+    fs.unlink(join(model_dir, f), (err) => {});
+  }
+  fs.rmdir(model_dir, (err) => {});
+  return true;
 }
 
 exports.lookup_fit = lookup_fit;
@@ -85,10 +93,13 @@ function store_fit(fit) {
 
 exports.delete_fit = delete_fit;
 function delete_fit(fit_id) {
-  // FIXME delete files
+  let fit = all_fits.get(fit_id);
+  if (fit === undefined)
+    return false;
   for (let [op_id, op] of all_ops)
     if (op.metadata.fit_id === fit_id)
       op.cancelled = true;
+  fs.unlink(join(cache_dir, fit.fit_file), (err) => {});
   return all_fits.delete(fit_id);
 }
